requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
